feat: Refactor database schema imports
