+++ src/server/db/schema/sessions.ts
@@ -5,7 +5,7 @@ import {
     text,
 } from "drizzle-orm/sqlite-core";
 
-import { createTable } from "..";
+import { createTable } from "../table";
 import { users } from "./users";
 
 export const sessions = createTable(
