+++ src/server/db/schema/status.ts
@@ -5,7 +5,7 @@ import {
     text,
 } from "drizzle-orm/sqlite-core";
 
-import { createTable } from "..";
+import { createTable } from "../table";
 import { protocolls } from "./protocoll";
 
 export const status = createTable(
