+++ src/server/db/schema/types.ts
@@ -5,7 +5,7 @@ import {
     index,
 } from "drizzle-orm/sqlite-core";
 
-import { createTable } from "..";
+import { createTable } from "../table";
 import { reports } from "./reports";
 
 export const types = createTable(
