+++ src/server/db/table.ts
@@ -1,0 +1,3 @@
+import { sqliteTableCreator } from "drizzle-orm/sqlite-core";
+
+export const createTable = sqliteTableCreator((name) => `fixmytown_${name}`);
